Separate bundler into a different type
